+++ src/edimeli/editor_melicus.py
@@ -5,7 +5,7 @@
 
 import sys, os, time, json, logging, re
 
-from ed_melicorum_utils import get_cfg_data, write_roman, write_roman_version
+from edimeli.utils import get_cfg_data, write_roman, write_roman_version
 from praedica_min import praedica_min
 from incoha import incoha
 from lectormelicus.lector_melicus import lege_tabulae_gabc, copy_conv_gabc_vars
