+++ src/edimeli/incoha.py
@@ -5,7 +5,7 @@ import sys, os
 import subprocess
 from pathlib import Path
 
-from ed_melicorum_utils import print_frame, get_cfg_data
+from edimeli.utils import print_frame, get_cfg_data
 
 
 # ---------------
