+++ src/edimeli/praedica_maj.py
@@ -4,7 +4,7 @@
 import os
 from pathlib import Path
 
-from ed_melicorum_utils import print_frame, get_cfg_data
+from edimeli.utils import print_frame, get_cfg_data
 
 # ---------------
 # CONFIGURATION
