+++ src/edimeli/scriptormelicus/scriptor_melicus.py
@@ -6,7 +6,7 @@
 import os
 from pathlib import Path
 
-from ed_melicorum_utils import print_frame, get_cfg_data
+from edimeli.utils import print_frame, get_cfg_data
 
 template_replace_map = {
     "TemplateTitle": "Title",
